# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lkdposts API",
    "version": "1.0.0",
    "description": "API responsável por disponibilizar funcionalidades de automação de posts.",
    "x-generated-at": "2025-09-25T16:37:38.453Z",
    "x-environment": "development"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "schemas": {
      "Envelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "format": "uuid",
                "example": "00000000-0000-4000-8000-000000000000"
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "INTERNAL_SERVER_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Internal server error"
              },
              "details": {
                "type": [
                  "object",
                  "null"
                ],
                "example": null
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string",
                "format": "uuid",
                "example": "00000000-0000-4000-8000-000000000000"
              }
            }
          }
        }
      },
...
      "FeedRefreshSummary": {
        "type": "object",
        "properties": {
          "feedId": {
            "type": "integer",
            "example": 42
          },
          "feedUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://news.example.com/rss"
          },
          "feedTitle": {
            "type": [
              "string",
              "null"
            ],
            "example": "News"
          },
          "skippedByCooldown": {
            "type": "boolean",
            "example": false
          },
          "cooldownSecondsRemaining": {
            "type": "integer",
            "example": 0
          },
          "itemsRead": {
            "type": "integer",
            "example": 25
          },
          "itemsWithinWindow": {
            "type": "integer",
            "example": 8
          },
          "articlesCreated": {
            "type": "integer",
            "example": 3
          },
          "duplicates": {
            "type": "integer",
            "example": 2
          },
          "invalidItems": {
            "type": "integer",
            "example": 1
          },
          "error": {
            "type": [
              "object",
              "null"
            ],
            "example": null,
            "properties": {
              "message": {
                "type": "string",
                "example": "Feed request timed out"
              }
            },
            "description": "Error returned when the feed fetch failed. Null when the refresh completed successfully."
          }
        }
      },
      "FeedRefreshResponse": {
        "type": "object",
        "properties": {
          "now": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-20T12:34:56.000Z"
          },
          "feeds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedRefreshSummary"
            }
          }
        }
      },
...
      "PostsCleanupResult": {
        "type": "object",
        "properties": {
          "removedArticles": {
            "type": "integer",
            "example": 12
          },
          "removedPosts": {
            "type": "integer",
            "example": 12
          }
        }
      },
      "PostFeedReference": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 7
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "example": "Tech Newsletter"
          },
          "url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri",
            "example": "https://tech.example.com/rss"
          }
        }
      },
      "PostContent": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
          },
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "example": "2025-01-20T12:35:30.000Z"
          }
        }
      },
      "PostListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 99
          },
          "title": {
            "type": "string",
            "example": "Latest tech news"
          },
          "contentSnippet": {
            "type": "string",
            "example": "Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-20T10:00:00.000Z"
          },
          "feed": {
            "type": [
              "object",
              "null"
            ],
            "allOf": [
              {
                "$ref": "#/components/schemas/PostFeedReference"
              }
            ],
            "example": {
              "id": 7,
              "title": "Tech Newsletter",
              "url": "https://tech.example.com/rss"
            }
          },
          "post": {
            "type": [
              "object",
              "null"
            ],
            "allOf": [
              {
                "$ref": "#/components/schemas/PostContent"
              }
            ],
            "example": {
              "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
              "createdAt": "2025-01-20T12:36:00.000Z"
            }
          }
        }
      },
      "PostListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostListItem"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lkd_session",
        "description": "Sessão autenticada emitida após login com Google."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Session token",
        "description": "Token de sessão transmitido via cabeçalho Authorization."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida (parâmetros ou payload incorretos).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "invalidInput": {
                "summary": "Valor inválido informado",
                "value": {
                  "success": false,
                  "error": {
                    "code": "INVALID_INPUT",
                    "message": "Invalid input data"
                  },
                  "meta": {
                    "requestId": "00000000-0000-4000-8000-000000000000"
                  }
                }
              },
              "urlRequired": {
                "summary": "URL obrigatória",
                "value": {
                  "success": false,
                  "error": {
                    "code": "URL_REQUIRED",
                    "message": "URL is required"
                  },
                  "meta": {
                    "requestId": "00000000-0000-4000-8000-000000000000"
                  }
                }
              },
              "invalidCursor": {
                "summary": "Cursor de paginação inválido",
                "value": {
                  "success": false,
                  "error": {
                    "code": "INVALID_CURSOR",
                    "message": "Invalid pagination cursor"
                  },
                  "meta": {
                    "requestId": "00000000-0000-4000-8000-000000000000"
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Autenticação requerida.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "success": false,
              "error": {
                "code": "UNAUTHENTICATED",
                "message": "Authentication required"
              },
              "meta": {
                "requestId": "00000000-0000-4000-8000-000000000000"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/posts/refresh": {
      "post": {
        "summary": "Refresh articles and generate posts from configured feeds",
        "description": "Consulta todos os feeds do usuário autenticado, aplica a janela de retenção configurada e cria posts placeholder para notícias recentes.",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Refresh executed for every feed owned by the user",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/FeedRefreshResponse"
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "now": "2025-01-20T12:34:56.000Z",
                        "feeds": [
                          {
                            "feedId": 1,
                            "feedUrl": "https://example.com/rss.xml",
                            "feedTitle": "Example RSS",
                            "skippedByCooldown": false,
                            "cooldownSecondsRemaining": 0,
                            "itemsRead": 12,
                            "itemsWithinWindow": 4,
                            "articlesCreated": 2,
                            "duplicates": 1,
                            "invalidItems": 0,
                            "error": null
                          },
                          {
                            "feedId": 2,
                            "feedUrl": "https://news.example.com/feed",
                            "feedTitle": null,
                            "skippedByCooldown": true,
                            "cooldownSecondsRemaining": 1800,
                            "itemsRead": 0,
                            "itemsWithinWindow": 0,
                            "articlesCreated": 0,
                            "duplicates": 0,
                            "invalidItems": 0,
                            "error": {
                              "message": "Feed request timed out"
                            }
                          }
                        ]
                      },
                      "meta": {
                        "requestId": "77777777-8888-4999-8aaa-bbbbbbbbbbbb"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/posts/cleanup": {
      "post": {
        "summary": "Delete old articles and posts outside the retention window",
        "description": "Remove artigos e posts vinculados aos feeds do usuário autenticado que estejam fora da janela de 7 dias utilizada pelo sistema.",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cleanup executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostsCleanupResult"
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "removedArticles": 6,
                        "removedPosts": 6
                      },
                      "meta": {
                        "requestId": "12345678-90ab-4cde-8fab-1234567890ab"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/posts": {
      "get": {
        "summary": "List generated posts derived from recent feed entries",
        "description": "Retorna artigos normalizados com o conteúdo de post correspondente, aplicando ordenação decrescente por data e paginação por cursor.",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Cursor de paginação retornado em chamadas anteriores."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Quantidade de itens por página (padrão 20, máximo 50)."
          },
          {
            "in": "query",
            "name": "feedId",
            "schema": {
              "type": "integer"
            },
            "description": "Filtra os artigos pertencentes ao feed informado."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of recent posts",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PostListResponse"
                        },
                        "meta": {
                          "type": "object",
                          "properties": {
                            "requestId": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "nextCursor": {
                              "type": [
                                "string",
                                "null"
                              ],
                              "description": "Cursor para a próxima página ou null ao final da lista."
                            },
                            "limit": {
                              "type": "integer",
                              "description": "Limite aplicado à consulta."
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "items": [
                          {
                            "id": 101,
                            "title": "Nova parceria anunciada",
                            "contentSnippet": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
                            "publishedAt": "2025-01-20T09:00:00.000Z",
                            "feed": {
                              "id": 1,
                              "title": "Example RSS",
                              "url": "https://example.com/rss.xml"
                            },
                            "post": {
                              "content": "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
                              "createdAt": "2025-01-20T12:35:00.000Z"
                            }
                          },
                          {
                            "id": 102,
                            "title": "Tech Weekly 42",
                            "contentSnippet": "Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
                            "publishedAt": "2025-01-19T17:30:00.000Z",
                            "feed": {
                              "id": 2,
                              "title": null,
                              "url": "https://news.example.com/feed"
                            },
                            "post": null
                          }
                        ]
                      },
                      "meta": {
                        "requestId": "55555555-6666-4777-8888-999999999999",
                        "nextCursor": "MjAyNS0wMS0xOVQxNzozMDowMC4wMDBaOjEwMg==",
                        "limit": 20
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
...
}